requirements.txt
django
djangorestframework
pymysql
django-redis